scripts/requirements.txt
pyswisseph>=2.10.3
pytz>=2024.1
